Remove all references to the combined parser
